+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+## Added
+
+- **! Added `Table.get_row_where(self: Self@Table, attribute_name: str, attribute_value: Any) -> tuple[Any, ...]`.**
+
 ## [0.3.0] - 2022-11-25
 
 ## Added